In this cut-and-pasted header, disambiguate sqrt and pull it from the std:: namespace as cmath doesn't have to do the using for you.
svn path=/trunk/koffice/; revision=838659

+++ krita/plugins/paintops/chalk/bristle.h
@@ -42,7 +42,7 @@ public:
 
 
 	float distanceCenter(){
-		return sqrt(m_x*m_x + m_y*m_y);
+		return std::sqrt((double)(m_x*m_x + m_y*m_y));
 	}
 	float amount();
 private:
